+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/domain/jwt/JwtUser.java
@@ -5,6 +5,7 @@ import io.tech1.framework.b2b.base.security.jwt.domain.identifiers.UserId;
 import io.tech1.framework.domain.base.Email;
 import io.tech1.framework.domain.base.Password;
 import io.tech1.framework.domain.base.Username;
+import io.tech1.framework.domain.tests.constants.TestsZoneIdsConstants;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -13,13 +14,17 @@ import java.time.ZoneId;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+
+import static io.tech1.framework.domain.base.AbstractAuthority.*;
+import static io.tech1.framework.domain.utilities.random.RandomUtility.*;
 
 public record JwtUser(
         UserId id,
         Username username,
         Password password,
         ZoneId zoneId,
-        List<SimpleGrantedAuthority> authorities,
+        Set<SimpleGrantedAuthority> authorities,
         Email email,
         String name,
         Map<String, Object> attributes
@@ -66,6 +71,37 @@ public record JwtUser(
                 this.username,
                 this.authorities,
                 this.zoneId
+        );
+    }
+
+    public static JwtUser random() {
+        return new JwtUser(
+                UserId.random(),
+                Username.random(),
+                Password.random(),
+                randomZoneId(),
+                Set.of(
+                        new SimpleGrantedAuthority(randomElement(List.of(SUPER_ADMIN, INVITATION_CODE_READ, INVITATION_CODE_WRITE)))
+                ),
+                Email.random(),
+                randomString(),
+                Map.of(
+                        randomString(), randomString(),
+                        randomString(), randomInteger()
+                )
+        );
+    }
+
+    public static JwtUser testsHardcoded() {
+        return new JwtUser(
+                UserId.testsHardcoded(),
+                Username.testsHardcoded(),
+                Password.testsHardcoded(),
+                TestsZoneIdsConstants.EET_ZONE_ID,
+                Set.of(),
+                Email.testsHardcoded(),
+                "",
+                Map.of()
         );
     }
 }