+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/sessions/SessionRegistry.java
@@ -22,7 +22,7 @@ public interface SessionRegistry {
     @Async
     void logout(Username username, JwtAccessToken accessToken);
 
-    // WARNING: think about migrating to separate service/registry
+    // think about migrating to separate service/registry
     void cleanByExpiredRefreshTokens(Set<Username> usernames);
     ResponseUserSessionsTable getSessionsTable(Username username, CookieAccessToken cookie);
 }