+++ tech1-framework-b2b-postgres-security-jwt/src/test-integration/java/io/tech1/framework/b2b/postgres/security/jwt/repositories/PostgresInvitationCodesRepositoryIT.java
@@ -14,18 +14,16 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.HashSet;
-import java.util.stream.Collectors;
 
 import static io.tech1.framework.b2b.base.security.jwt.comparators.SecurityJwtSorts.INVITATION_CODES_UNUSED;
 import static io.tech1.framework.b2b.base.security.jwt.constants.SecurityJwtConstants.DEFAULT_INVITATION_CODE_LENGTH;
 import static io.tech1.framework.b2b.base.security.jwt.tests.random.BaseSecurityJwtRandomUtility.randomInvitationCode;
+import static io.tech1.framework.b2b.base.security.jwt.utilities.SpringAuthoritiesUtility.getSimpleGrantedAuthorities;
 import static io.tech1.framework.b2b.postgres.security.jwt.tests.random.PostgresSecurityJwtDbDummies.dummyInvitationCodesData1;
 import static io.tech1.framework.b2b.postgres.security.jwt.tests.random.PostgresSecurityJwtDbDummies.dummyInvitationCodesData2;
-import static io.tech1.framework.domain.tests.constants.TestsUsernamesConstants.TECH1;
 import static io.tech1.framework.domain.utilities.random.EntityUtility.entity;
 import static io.tech1.framework.domain.utilities.random.RandomUtility.*;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -160,13 +158,13 @@ class PostgresInvitationCodesRepositoryIT extends TestsApplicationRepositoriesRu
         assertThat(this.invitationCodesRepository.isPresent(notExistentInvitationCodeId).present()).isFalse();
 
         // Act-Assert-3
-        this.invitationCodesRepository.saveAs(TECH1, requestNewInvitationCodeParams);
+        this.invitationCodesRepository.saveAs(Username.testsHardcoded(), requestNewInvitationCodeParams);
         assertThat(this.invitationCodesRepository.count()).isEqualTo(8);
-        var ownedInvitationCodes = this.invitationCodesRepository.findByOwner(TECH1);
+        var ownedInvitationCodes = this.invitationCodesRepository.findByOwner(Username.testsHardcoded());
         assertThat(ownedInvitationCodes).hasSize(1);
         var ownedInvitationCode = ownedInvitationCodes.get(0);
-        assertThat(ownedInvitationCode.getOwner()).isEqualTo(TECH1);
-        assertThat(ownedInvitationCode.getAuthorities()).isEqualTo(requestNewInvitationCodeParams.authorities().stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList()));
+        assertThat(ownedInvitationCode.getOwner()).isEqualTo(Username.testsHardcoded());
+        assertThat(ownedInvitationCode.getAuthorities()).isEqualTo(getSimpleGrantedAuthorities(requestNewInvitationCodeParams.authorities()));
         assertThat(ownedInvitationCode.getValue()).hasSize(40);
     }
 }