+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/crons/AbstractCronAction.java
@@ -1,6 +1,5 @@
 package io.tech1.framework.domain.crons;
 
-// [migrate]: tech1-framework
 @FunctionalInterface
 public interface AbstractCronAction {
     void execute() throws Exception;