+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/hardware/memories/SystemMemories.java
@@ -4,4 +4,11 @@ public record SystemMemories(
         GlobalMemory global,
         CpuMemory cpu
 ) {
+
+    public static SystemMemories testsHardcoded() {
+        return new SystemMemories(
+                GlobalMemory.testsHardcoded(),
+                CpuMemory.testsHardcoded()
+        );
+    }
 }