+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/tests/constants/TestsJunitConstants.java
@@ -4,6 +4,9 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class TestsJunitConstants {
+    public static final int TWICE = 2;
+    public static final int FIVE_TIMES = 5;
+
     public static final int SMALL_ITERATIONS_COUNT = 10;
     public static final int RANDOM_ITERATIONS_COUNT = 100;
 }