+++ tech1-framework-domain/src/test/java/io/tech1/framework/domain/tests/domain/enums/TestAuthority.java
@@ -2,7 +2,7 @@ package io.tech1.framework.domain.tests.domain.enums;
 
 import io.tech1.framework.domain.base.AbstractAuthority;
 
-// WARNING: used in PropertiesAsserterAndPrinterTest (reflection scan process)
+// Used in PropertiesAsserterAndPrinterTest (reflection scan process)
 public enum TestAuthority implements AbstractAuthority {
     USER("user"),
     ADMIN("admin");