+++ tech1-framework-properties/src/main/java/io/tech1/framework/properties/ApplicationFrameworkProperties.java
@@ -21,7 +21,6 @@ public class ApplicationFrameworkProperties implements PriorityOrdered {
     private EmailConfigs emailConfigs;
     private IncidentConfigs incidentConfigs;
     private HardwareMonitoringConfigs hardwareMonitoringConfigs;
-    // TODO [YL] delete legacy configs
     private HardwareServerConfigs hardwareServerConfigs;
     // b2b-base
     private SecurityJwtConfigs securityJwtConfigs;